Signed-off-by: Junhao Wang <wjunhao@amazon.com>

## What does this PR do?

This PR introduces a new edit-entity undo node for override editing. This node is a more selective version of avoid comparing instance DOMs to generate patches. As a result, it would increase the editor performance for override editing as mentioned in https://github.com/o3de/o3de/issues/11401.

This PR depends on https://github.com/o3de/o3de/pull/13640 to add override subtree successfully.

| Performance Improvement | Base | This PR | Change % |
| :---    |  ---: | ---: | ---: |
| Edit entity in 10k-entity prefab (Single-Edit) |  4,000 ms |  2,700 ms | -32.5% |
| Toggle lock on 1k-entity prefab (Multi-Edit) |  56,000 ms |  8,000 ms | -85.7% |

#### Major Changes
1. New undo node for edit-entity override that supports batch processing, merging override patches to add-entity patch. It can even be used for add-entity override workflow in the future.
2. Add patch index counter and functionality to generate subtree in Link class.
3. Update edit-entity workflow to partially adopt this undo node

#### What's Next

1. Further improvement in edit-entity workflow:
    1. Batch processing dirty entities in edit-entity workflow (improvement is noted in the GitHub issue)
    1. Can use this undo node for reparent editing as well
1. Can use this undo node for add-entity workflow (improvement is not obvious though)

## How was this PR tested?

- [x] Manually testing in editor.
- [x] Passed all tests.
- [x] Ran performance check. See improvement.
- [ ] Add some new unit tests for the new undo node